Filter runtime API metadata by version implemented
