+++ substrate/primitives/api/proc-macro/src/runtime_metadata.rs
@@ -15,13 +15,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::BTreeMap;
 use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
-use syn::{parse_quote, ItemImpl, ItemTrait, Result};
+use syn::{parse_quote, spanned::Spanned, ItemImpl, ItemTrait, Result};
 
 use crate::{
 	common::CHANGED_IN_ATTRIBUTE,
 	utils::{
+		extract_api_version,
 		extract_impl_trait, filter_cfg_attributes, generate_crate_access,
 		generate_runtime_mod_name_for_trait, get_doc_literals, RequireQualifiedTraitPath,
 	},
@@ -72,7 +74,7 @@ fn collect_docs(attrs: &[syn::Attribute], crate_: &TokenStream2) -> TokenStream2
 ///
 /// The metadata is exposed as a generic function on the hidden module
 /// of the trait generated by the `decl_runtime_apis`.
-pub fn generate_decl_runtime_metadata(decl: &ItemTrait) -> TokenStream2 {
+pub fn generate_decl_runtime_metadata(decl: &ItemTrait, method_versions: &BTreeMap<&syn::Ident, u64>) -> TokenStream2 {
 	let crate_ = generate_crate_access();
 	let mut methods = Vec::new();
 
@@ -127,6 +129,7 @@ pub fn generate_decl_runtime_metadata(decl: &ItemTrait) -> TokenStream2 {
 
 		// String method name including quotes for constructing `v15::RuntimeApiMethodMetadata`.
 		let method_name = signature.ident.to_string();
+		let method_version = method_versions.get(&signature.ident).copied().unwrap_or(1);
 		let docs = collect_docs(&method.attrs, &crate_);
 
 		// Include the method metadata only if its `cfg` features are enabled.
@@ -139,6 +142,7 @@ pub fn generate_decl_runtime_metadata(decl: &ItemTrait) -> TokenStream2 {
 			#( #attrs )*
 			#crate_::metadata_ir::RuntimeApiMethodMetadataIR {
 				name: #method_name,
+				version: #method_version,
 				inputs: #crate_::vec![ #( #inputs, )* ],
 				output: #output,
 				docs: #docs,
@@ -242,10 +246,35 @@ pub fn generate_impl_runtime_metadata(impls: &[ItemImpl]) -> Result<TokenStream2
 			*segment = parse_quote!(#mod_name);
 		}
 
+		// We want to filter runtime API metadata to only include the methods that we are actually
+		// implementing. So, get the `api_version` that we are implementing (1 by default),
+		// handling the possibility that it's guarded by a #[cfg_attr] attribute, and generate a
+		// relevant filter method body based on this.
+		let api_version_comparison = {
+			let api_version = extract_api_version(&impl_.attrs, impl_.span())?;
+			let base_version = api_version.custom.unwrap_or(1);
+			if let Some(cfg_version) = api_version.feature_gated {
+				let cfg_feature = cfg_version.0;
+				let cfg_version = cfg_version.1;
+				quote! {
+					#[cfg(feature = #cfg_feature)]
+					{ method.version <= #cfg_version }
+					#[cfg(not(feature = #cfg_feature))]
+					{ method.version <= #base_version }
+				}
+			} else {
+				quote! {
+					method.version <= #base_version
+				}
+			}
+		};
+
 		let attrs = filter_cfg_attributes(&impl_.attrs);
 		metadata.push(quote!(
 			#( #attrs )*
-			#trait_::runtime_metadata::#generics()
+			#trait_::runtime_metadata::#generics().keeping_methods(|method| {
+				#api_version_comparison
+			})
 		));
 	}
